Add files via upload
Convolutional Neural Networks Project: Street View Housing Number Digit Recognition

Welcome to the project on classification using Convolutional Neural Networks. I am going to continue to work with the Street View Housing Numbers (SVHN) image dataset for this project.
Context: One of the most interesting tasks in deep learning is to recognize objects in natural scenes. The ability to process visual information using machine learning algorithms can be very useful as demonstrated in various applications.
The SVHN dataset contains over 600,000 labeled digits cropped from street-level photos. It is one of the most popular image recognition datasets. It has been used in neural networks created by Google to improve the map quality by automatically transcribing the address numbers from a patch of pixels. The transcribed number with a known street address helps pinpoint the location of the building it represents.
Objective: To build a CNN model that can recognize the digits in the images.
Dataset: Here, we will use a subset of the original data to save some computation time. The dataset is provided as a .h5 file. The basic preprocessing steps have been applied on the dataset.